+++ client/js/modules/uptime.js
@@ -4,16 +4,16 @@ export const processUptime = (data) => {
     const {position, rate, score} = data
     let color = "neutral", icon = "infinite"
 
-    if (Metro.utils.between(position, 0, 75)) {
-        color = 'success'
-        icon = 'checkmark'
-    } else if (Metro.utils.between(position, 75, 100, true)) {
-        color = 'warning'
-        icon = 'warning'
-    } else if (Metro.utils.between(position, 101, 120, true)) {
-        color = 'alert'
-        icon = 'bin'
-    }
+    // if (Metro.utils.between(position, 0, 75)) {
+    //     color = 'success'
+    //     icon = 'checkmark'
+    // } else if (Metro.utils.between(position, 75, 100, true)) {
+    //     color = 'warning'
+    //     icon = 'warning'
+    // } else if (Metro.utils.between(position, 101, 120, true)) {
+    //     color = 'alert'
+    //     icon = 'bin'
+    // }
 
     $("#sidecar-position").text(position).removeClassBy("label-").addClass(`label-${color}`)
     $("#position-icon").removeClassBy("label-").removeClassBy("mif-").addClass(`label-${color}`).addClass(`mif-${icon}`)